// src/app/models/log-digitalizado.ts
import { Actividad } from "src/app/models";

export class LogDigitalizado {
  idLog: number;
  suministro: number;
  actividad: Actividad;
  ordTrabOrdServCedu: string;
  tipologia: string;
  tipoArchivo: string;
  usuarioAccion: string;
  fechaHoraAccion: Date;
  tipoAccion: string;
  ipAccion: string;
  }